refactor: Modularize Cloudflare and logging functionality into separate modules
This commit moves the Cloudflare and logging related code into their own dedicated modules. This separation enhances code organization and readability, making it easier to maintain and understand the codebase. The Cloudflare module now handles all Cloudflare related operations, and the logging module manages all logging functionality.
